requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==5.0.1
billiard==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
dataclass-wizard==0.22.3
django==4.2.23
django-cas-ng==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-taggit==6.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.6
kombu==5.3.5
lxml==6.0.0
mailhog==0.1.0
packaging==25.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.10
pypng==0.20220715.0
python-cas==1.6.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
